polish publish view
